Moved components to their own folder
